Add a function with a stable to environmental event stream packages to convert stream to iterable
