+++ packages/eventstream-chunker-browser/src/iterableForEventStream.ts
@@ -1,0 +1,8 @@
+import { BlobEventStreamSplitter } from './BlobEventStreamSplitter';
+import { Message } from '@aws/types';
+
+export function iterableForEventStream(
+    blob: Blob
+): AsyncIterableIterator<Message> {
+    return new BlobEventStreamSplitter(blob);
+}
